Fixed drop course page link bug
